requirements.txt
streamlit==1.39.0
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.30.0
PyMuPDF==1.23.26
python-docx==1.0.1
